correct location of fonts
Since the fonts/ directory is not inside the css/ directory, you have to change to its parent directory first before you go to fonts. To do that, you need to add two dots [..] to tell the browser that this relative location of the font is located to the stylesheet's parent directory and to another folder. Hence:

../fonts/Poppins-Regular.ttf
../fonts/Poppins-ExtraBold.ttf

When you write `fonts/Poppins-Regular.ttf`, you are telling the browser to look for a fonts/ directory inside the css/ directory.